feat: implement SessionHandler everywhere, set all Ids to be identities, await all DB writes and reads, return Id from inserts

+++ CS.Session.Infrastructure/State/SessionStateHandler.cs
@@ -8,18 +8,27 @@ namespace CS.Session.Infrastructure.State
     public class SessionStateHandler : ISessionStateHandler
     {
         private readonly RedisCacheService _cacheService;
-        public SessionStateHandler(RedisCacheService cacheService)
+        private readonly ISessionHandler _sessionHandler;
+
+        public SessionStateHandler(RedisCacheService cacheService, ISessionHandler sessionHandler)
         {
             _cacheService = cacheService;
+            _sessionHandler = sessionHandler;
         }
 
-        public async Task OnStateClosed(string userIP, string sessionState)
+        public async Task OnStateClosed(string sessionIdString, string userIP, string sessionState)
         {
             if (sessionState != SessionState.CLOSED.ToString()) return;
+            if (!long.TryParse(sessionIdString, out long sessionId)) throw new Exception("SessionId is not a valid long number!");
+
+            // You could also return lastPingTimestamp from the Job / JobStateFilter
+            // to be more precise
+            DateTime lastPingTimestamp = DateTime.Now;
+
+            await _sessionHandler.UpdateSessionState(sessionId, SessionState.CLOSED, lastPingTimestamp);
 
             var redisKey = SessionUtil.GetRedisKey(userIP);
             await _cacheService.DeleteAsync(redisKey);
-            // TODO: PUT Session.CLOSED to DB
         }
     }
 }
